feat: complete user registration

+++ src/utils/cloudinary.js
@@ -1,27 +1,39 @@
-import { v2 as cloudinary } from 'cloudinary';
+import dotenv from "dotenv";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-    // Configuration
-    cloudinary.config({ 
-        cloud_name: process.env.COLUDINARY_CLOUD_NAME, 
-        api_key: process.env.COLUDINARY_API_KEY, 
-        api_secret:process.env.COLUDINARY_API_SECRET
-    });
+dotenv.config({
+    path: "./.env"
+});
 
- // Upload an file
-     const uploadOnCloudinary = async (locaFilePath) => {
-        try {
-            if (!locaFilePath) return null
-            
-           const response = await cloudinary.uploader.upload(locaFilePath,{
-                resource_type: "auto"
-            })
-            console.log("file is uploaded on cloudinary",response.url);
-            return response;
-        } catch (error) {
-            fs.unlinkSync(localFilePath)
-            return null;
+// Configuration
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
+});
+
+// Upload a file
+const uploadOnCloudinary = async (localFilePath) => {
+    try {
+        if (!localFilePath) return null;
+
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto"
+        });
+
+        console.log("file is uploaded on cloudinary", response.url);
+
+        return response;
+    } catch (error) {
+        console.log("CLOUDINARY ERROR:", error);
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
         }
-     }
-      
-    export {cloudinary, uploadOnCloudinary}
+
+        return null;
+    }
+};
+
+export { cloudinary, uploadOnCloudinary };
